Refactor misc code
